aspeed/ast2500: Add SDRAM MC driver
The driver is very ast2500 specific and is completely incompatible
with previous versions of the chip.

The memory controller is very poorly documented by Aspeed in the
datasheet, with any mention of the whole range of registers missing. The
initialization procedure has been basically taken from Aspeed SDK, where
it is implemented in assembly and rewritten in C, with very limited
understanding of what exacly it is doing.